+++ src/helpers/constants.helper.js
@@ -1,0 +1,8 @@
+const CACHE_TTL = {
+  MANGA_LIST: 5 * 60 * 1000,
+  MANGA_DETAIL: 5 * 60 * 1000,
+  GENRES: 60 * 60 * 1000,
+  STATUS: 24 * 60 * 60 * 1000,
+};
+
+module.exports = { CACHE_TTL };